C_flutter/python/requirements.txt
PyMuPDF
pytesseract
Pillow
beautifulsoup4
lxml
html5lib